Set up individiual chat rooms
The logic for this change is centralised around keeping the room details
in room.py, and websocket connection handlers are now associated to
a particular room.

The client code templates still needs to be rewritten to reflect this
change since the index page sends people to a random, empty room by
default.