Replace database logic with redux-like framework
Persistent state has been removed and the logic of the code is beginning
to be moved into the singular data store, which will probably get
refactored out eventually.  The architecture is based on redux
(javascript) and re-frame (clojure), which are both unidirectional
application architectures, although they are for single page apps rather
than the multi-page website being designed in web500.